+++ generators/provider/inputs/fileName.js
@@ -1,11 +1,11 @@
 /**
  * Generated header by C0ZEN for generator-cozen-angular project
- * Generated file providerPath on WebStorm
+ * Generated file fileName on WebStorm
  *
  * Created by: Geoffrey "C0ZEN" Testelin
  * Date: 28/12/2017
- * Time: 14:15
- * Version: 0.8.0
+ * Time: 19:07
+ * Version: 1.0.0
  */
 (function () {
 	'use strict';
@@ -15,20 +15,23 @@
 	const prompts = [
 		{
 			type   : 'input',
-			name   : 'providerPath',
-			message: 'Path of the provider :',
-			default: 'scripts/providers'
+			name   : 'fileName',
+			message: 'Name of file :'
 		}
 	];
 
 	const hints = [
-		'This will allow you to override the default path for the provider'
+		'This is the name of the file - only the prefix - in camelCase.',
+		'We will automatically add the type and the extension as <fileName>.<ttype>.<extension>',
+		'Example: myFileName.provider.js'
 	];
 
-	module.exports = class ProviderPath {
+	module.exports = class ProviderName {
 		static set($that) {
 			logs.hintHeader($that);
 			logs.hintsMultiline($that, hints);
+
+			prompts[0].default = $that.providerName;
 
 			return $that.prompt(prompts).then($response => {
 				this.onSuccess($that, $response);
@@ -36,8 +39,8 @@
 		}
 
 		static onSuccess($that, $response) {
-			$that.providerPath = $response.providerPath;
-			$that.config.set('providerPath', $that.providerPath);
+			$that.fileName = $response.fileName;
+			$that.config.set('fileName', $that.fileName);
 			$that.log();
 		}
 	};