+++ CHANGELOG.md
@@ -5,11 +5,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- New input `fileName` for the provider to customize the name of the file
+
 ### Changed
 - Partial header to add the name of the generator
+- Added the default path as default value for the provider path
 
 ### Fixed
 - Error with wrong variable in the provider template
+
+### Removed
+- Default callback path for the provider when the custom path is empty
 
 ## [0.9.0]
 ### Added